gabomazone-app/requirements.txt
# ── Framework principal
Django>=5.1,<5.2
gunicorn>=21.0.0

# ── Base de données
psycopg2-binary>=2.9.9

# ── Cache & Queue
redis>=5.0.0
celery>=5.3.6
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# ── Configuration
python-decouple>=3.8

# ── Sécurité
django-ratelimit>=4.1.0

# ── Assets statiques
whitenoise>=6.6.0

# ── Images
Pillow>=10.0.0

# ── Rich text editor
django-ckeditor>=6.7.0
django-js-asset>=2.0.0

# ── Formulaires
django-crispy-forms>=2.1
crispy-bootstrap4>=2023.1
django-bootstrap4>=23.4

# ── Géo / pays
django-countries>=7.5.1

# ── Monnaie
django-currencies>=0.10.1

# ── Email
django-sendgrid-v5>=1.2.0
sendgrid>=6.11.0

# ── Admin theme (django-adminlte3 inclut déjà adminlte3_theme)
django-adminlte3==0.1.6

# ── Captcha
django-simple-captcha>=0.5.20
django-recaptcha>=4.0.0

# ── PDF
weasyprint>=60.0

# ── Internationalisation
pytz>=2023.3

# ── Utilitaires
requests>=2.31.0
beautifulsoup4>=4.12.0
python-stdnum>=1.19
django-jsonfield-backport>=1.0.5
setuptools>=70.0.0

# ── Monitoring
django-user-agents>=0.4.0
ua-parser>=0.18.0
user-agents>=2.2.0

# ── Localisation
django-localflavor>=3.1
